chore: update metadata.yaml update script to be compatible with Ship.js (#48)
I discovered that Ship.js runs inside an ancient, no longer updated,
library called [esm](https://github.com/standard-things/esm). The YAML
library uses some modern features in it, especially the optional
chaining operator (`?.`) which the [esm library is incompatible
with](https://github.com/standard-things/esm/issues/866), and as such we
need to execute it out of band until Ship.js stops using that library
(it's not necessary any more as ESM support is in Node itself, but it's
not a trivial update to modernise Ship.js to not require that — I
started down that route… not fun).